Massive refactoring
