+++ src/editors/protocol104/connectedap-editor.ts
@@ -10,7 +10,7 @@ import '@material/mwc-fab';
 
 import '../../action-icon.js';
 import { newWizardEvent, newActionEvent } from '../../foundation.js';
-import { editConnectedApWizard } from './wizards/connectedap.js';
+import { editConnectedAp104Wizard } from '../../wizards/connectedap.js';
 
 /** [[`104`]] subeditor for a `ConnectedAP` element. */
 @customElement('connectedap-editor')
@@ -25,7 +25,7 @@ export class ConnectedAPEditor extends LitElement {
   }
 
   private openEditWizard(): void {
-    this.dispatchEvent(newWizardEvent(editConnectedApWizard(this.element)));
+    this.dispatchEvent(newWizardEvent(editConnectedAp104Wizard(this.element)));
   }
 
   remove(): void {
